+++ jeap-process-archive-adapter-db/src/main/resources/db/migration/V3__make_backfill_reference_version_nullable.sql
@@ -1,0 +1,2 @@
+alter table backfill_task
+    alter column reference_version drop not null;