+++ jeap-process-archive-adapter-rest-api/src/main/java/ch/admin/bit/jeap/processarchive/adapter/restapi/BackfillJobRequest.java
@@ -3,7 +3,6 @@ package ch.admin.bit.jeap.processarchive.adapter.restapi;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
-import jakarta.validation.constraints.NotNull;
 
 import java.util.List;
 
@@ -12,6 +11,6 @@ public record BackfillJobRequest(
         @NotBlank String topic,
         @NotEmpty List<@Valid ArchiveDataReferenceRequest> archiveDataReferences) {
 
-    public record ArchiveDataReferenceRequest(@NotBlank String id, @NotNull Integer version) {
+    public record ArchiveDataReferenceRequest(@NotBlank String id, Integer version) {
     }
 }