+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [14.1.2] - 2026-06-23
+
+### Added
+- Add documentation for backfill jobs
+
+### Changed
+- Make the backfill reference version optional
+- Use collision-safe idempotence ids for backfill commands with optional reference versions
+
 ## [14.1.1] - 2026-06-18
 
 ### Added